backend/requirements.txt
# Core
fastapi==0.115.8
uvicorn[standard]==0.34.0
python-multipart==0.0.20

# Database
sqlalchemy==2.0.38
alembic==1.14.1
psycopg2-binary==2.9.10
asyncpg==0.30.0

# Auth & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
cryptography==44.0.0

# Settings
pydantic-settings==2.8.1
pydantic[email]==2.10.6
python-dotenv==1.0.1

# Cache & Task Queue
redis==5.2.1
celery[redis]==5.4.0
celery[beat]==5.4.0
flower==2.0.1

# AI
openai==1.63.2

# Google APIs (YouTube)
google-auth==2.38.0
google-auth-oauthlib==1.2.1
google-api-python-client==2.161.0

# Billing
stripe==11.4.1
mercadopago==2.3.0
requests==2.32.3

# Email
resend==2.7.0

# Utils
httpx==0.28.1
python-slugify==8.0.4
Pillow==11.1.0

# Dev / Test
pytest==8.3.4
pytest-asyncio==0.25.3
httpx==0.28.1
ruff==0.9.7
